admin theme

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,7 +19,7 @@ class CreateUsersTable extends Migration
             $table->string('email')->nullable();
             $table->string('phone_number', 15)->nullable();
             $table->string('password');
-            $table->bigInteger('user_role_id')->default(0);
+            $table->integer('user_role_id')->default(0);
             $table->tinyInteger('is_block_admin')->default(0);
             $table->timestamp('email_verified_at')->nullable();
             $table->string('reset_code')->nullable();
